#21 refactor: mypage controller를 member controller에 통합
